move Disclosure  to its own file
